PRC-577: Contact search sort fixes & improvements (#65)
Made date of birth sort with nulls as if they are the eldest so nulls first/last depending on asc/desc.

Fixed issue where if there are multiple fields being sorted on it was only the last applied.

Removed default sort which was undocumented and might provide unexpected results.

Properly map all fields from API to db type. This is to make things like sorting on id work where the field in the API is called `id` but in the database entity it's called `contactId` so a user would sort on `id` as that's what's in the API but it wouldn't work and failed silently.