PRC-578: Prisoner contact list improvements (#66)
* PRC-578: Expand filtering and sorting for prisoner contact list

Fix the sorting so that it works the same as contact search. DOB with nulls meaning eldest first and mapping API to entity field names.

Add ability to filter
* active, inactive or both (default)
* social, official or both (default)
* emergency contact or not or either (default)
* next of kind or not or either (default)
* emergency contact or next of kin, not EC and not NOK, ignore EC & NOK (default)

* PRC-578: Add restriction summary to contact list

Create a combined summary of global and prisoner-contact restrictions. The idea is to be able to quickly show which restrictions currently apply to a prisoners contact.