PRC-136: Page and sort linked prisoners (#68)
Changed linked prisoners to page in memory to allow sorting based on name then prisoner number. I've manually tested this with 10000 linked prisoners and it took less than 15 seconds.

Also move to PagedModel instead of PageImpl to fix a spring warning about using an unsafe type. This is a breaking change and needs some fixes to the client code.

Also fixed a number of swagger issues with the paged endpoints.