PRC-725 & PRC-696: Reject duplicate relationships (#125)
* PRC-725 & PRC-696: Reject duplicate relationships

Reject duplicate relationships (same prisoner number, contact id and relationship to prisoner code) when a new relationship is created by sync or DPS API. Also reject duplicate relationships when updating a contact relationship but only if the user is changing the relationship. Do not reject sync updates to relationships.

This allows existing duplicates in NOMIS to be updated but blocks the creation of new duplicates, we trust NOMIS not to create any more new duplicates as it is validated in the NOMIS UI now.

* test name

* Add missing swagger doc for prisoner contact conflict