PRC-746: Allow sync to create dupes if not current term (#133)
Sometimes users do weird things like create a contact and then create the same contact on an old booking for seemingly no reason. When currentTerm=false skip the duplicate check as we only care about current duplicates.

There is an edge case where a dupe is created on old bookings and then the old booking is reactivated but we feel this is the safest overall option.