PRC-797: Support existing relationships param on search (#172)
A new optional query parameter which takes a prisoner number and checks if any of the matching contacts within contact search already have any relationship to the prisoner and then returns those existing relationships if it does.
This is to support alerting users that they might be about to create a duplicate contact relationship for a prisoner within the search results.

Also split the search function out of ContactService into a new service just for searching as the number of dependencies was getting out of hand as was the test.

Moved doc and validation of ContactSearchRequest out of the request object into the controller as it does not render the descriptions in swagger (not sure if it ever did but I couldn't get it working)